[Feat] 커뮤니티탭 맛짱투표

+++ community/domain/src/main/java/com/threedollar/domain/data/PollItem.kt
@@ -37,7 +37,7 @@ data class PollItem(
         ) {
             data class Choice(
                 val count: Int, // 0
-                val ratio: Int, // 0
+                val ratio: Double, // 0
                 val selectedByMe: Boolean // true
             )
         }
